migrate create-entity page
